+++ doc/things-learned.md
@@ -1,3 +1,5 @@
 # Things Learned
 
 component mounting and lifecycle
+
+The React property *Key* is internal and only accesible by React
