+++ src/atoms/Buttons.js
@@ -1,0 +1,5 @@
+import React, { Component } from 'react';
+
+class Button extends Component {
+  
+}
